Fix: Avoid over-purging dependencies in skip-ocaml
The skip-ocaml algorithm handles dependency cleanup differently from
the usual Skip runtime. In Skip, all data lives in a context and is
naturally garbage-collected if a crash occurs. However, in skip-ocaml,
we manually clean up to avoid leaking dependencies in the event of a
crash.

Unfortunately, the logic for clearing dependencies was too aggressive,
causing even basic examples to break. For now, we’ll prioritize
correctness and accept the possibility of leaking dependencies on
crash. We can revisit cleanup strategies later.